requirements.txt
networkx==2.6.2
numpy==1.20.2
PyYAML==5.4.1
scipy==1.7.0
scikit-learn==0.24.2
torch==1.6.0+cu101
